renamed ErrorPage to EmptyState and corrected the imports
