+++ src/pages/index.js
@@ -1,8 +1,8 @@
 import Cart from "./Cart/Cart";
-import Wishlist from "./WishList/WishList";
+import WishList from "./WishList/WishList";
 import Home from "./Home/Home";
 import Product from "./Product/Product";
 import ProductList from "./ProductList/ProductList";
 import EmptyState from "./EmptyState/EmptyState";
 
-export { Cart, Wishlist, Home, Product, ProductList, EmptyState };
+export { Cart, WishList, Home, Product, ProductList, EmptyState };
